+++ app/dao/jobs_dao.py
@@ -4,15 +4,13 @@ from datetime import datetime, timedelta
 from flask import current_app
 from notifications_utils.statsd_decorators import statsd
 from sqlalchemy import (
-    Date as sql_date,
     asc,
-    cast,
     desc,
     func,
 )
 
 from app import db
-from app.dao import days_ago
+from app.utils import days_ago
 from app.models import (
     Job,
     JOB_STATUS_PENDING,
@@ -53,7 +51,7 @@ def dao_get_jobs_by_service_id(service_id, limit_days=None, page=1, page_size=50
         Job.original_file_name != current_app.config['ONE_OFF_MESSAGE_FILENAME'],
     ]
     if limit_days is not None:
-        query_filter.append(cast(Job.created_at, sql_date) >= days_ago(limit_days))
+        query_filter.append(Job.created_at >= days_ago(limit_days))
     if statuses is not None and statuses != ['']:
         query_filter.append(
             Job.job_status.in_(statuses)
